requirements.txt
folium==0.14.0
numpy==1.23.5
pandas==1.5.3
Pillow==9.4.0
Pillow==10.0.0
plotly==5.9.0
streamlit==1.25.0
streamlit_folium==0.13.0
